Issue 42: Setup main recipe list UI to use new data model.
* Moved Recipe model class to a separate models folder and file.
* Changed main recipe list to sue new data snapshot structure.
