+++ .flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"cloud_firestore","path":"/Users/shrop/bin/flutter/.pub-cache/hosted/pub.dartlang.org/cloud_firestore-0.12.9+5/","dependencies":["firebase_core"]},{"name":"firebase_core","path":"/Users/shrop/bin/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_core-0.4.0+9/","dependencies":[]}],"android":[{"name":"cloud_firestore","path":"/Users/shrop/bin/flutter/.pub-cache/hosted/pub.dartlang.org/cloud_firestore-0.12.9+5/","dependencies":["firebase_core"]},{"name":"firebase_core","path":"/Users/shrop/bin/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_core-0.4.0+9/","dependencies":[]}],"macos":[],"linux":[],"windows":[],"web":[]},"dependencyGraph":[{"name":"cloud_firestore","dependencies":["firebase_core"]},{"name":"firebase_core","dependencies":[]}],"date_created":"2020-05-18 08:14:14.536838","version":"1.17.1"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"cloud_firestore","path":"/Users/shrop/bin/flutter/.pub-cache/hosted/pub.dartlang.org/cloud_firestore-0.12.9+5/","dependencies":["firebase_core"]},{"name":"firebase_core","path":"/Users/shrop/bin/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_core-0.4.0+9/","dependencies":[]}],"android":[{"name":"cloud_firestore","path":"/Users/shrop/bin/flutter/.pub-cache/hosted/pub.dartlang.org/cloud_firestore-0.12.9+5/","dependencies":["firebase_core"]},{"name":"firebase_core","path":"/Users/shrop/bin/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_core-0.4.0+9/","dependencies":[]}],"macos":[],"linux":[],"windows":[],"web":[]},"dependencyGraph":[{"name":"cloud_firestore","dependencies":["firebase_core"]},{"name":"firebase_core","dependencies":[]}],"date_created":"2020-05-19 22:51:53.460264","version":"1.17.1"}
